move all TAU validation function to goods/helpers
