+++ applications/helpers.py
@@ -1,5 +1,3 @@
-from rest_framework.exceptions import ValidationError
-
 from applications.enums import ApplicationExportType
 from applications.models import BaseApplication
 from applications.serializers.end_use_details import (
@@ -40,7 +38,6 @@ from applications.serializers.standard_application import (
 from applications.serializers.temporary_export_details import TemporaryExportDetailsUpdateSerializer
 from cases.enums import CaseTypeSubTypeEnum, CaseTypeEnum
 from conf.exceptions import BadRequestError
-from goods.enums import Component
 from lite_content.lite_api import strings
 
 
@@ -129,51 +126,4 @@ def get_temp_export_details_update_serializer(export_type):
     else:
         raise BadRequestError(
             {f"get_temp_export_details_update_serializer does " f"not support this export type: {export_type}"}
-        )
-
-
-def validate_good_component_details(data):
-    """ Validate the accompanying details for the chosen 'yes' component option. """
-    component = data["is_component"]
-    component_detail_options = {
-        Component.YES_DESIGNED: {
-            "details_field": "designed_details",
-            "error": strings.Goods.NO_DESIGN_COMPONENT_DETAILS,
-        },
-        Component.YES_MODIFIED: {
-            "details_field": "modified_details",
-            "error": strings.Goods.NO_MODIFIED_COMPONENT_DETAILS,
-        },
-        Component.YES_GENERAL_PURPOSE: {
-            "details_field": "general_details",
-            "error": strings.Goods.NO_GENERAL_COMPONENT_DETAILS,
-        },
-    }
-
-    field = component_detail_options[component]["details_field"]
-    error = component_detail_options[component]["error"]
-
-    if not data[field]:
-        return {"is_valid": False, "details_field": field, "error": error}
-    else:
-        return {"is_valid": True, "details_field": field}
-
-
-def validate_component_fields(data):
-    if data.get("is_component_step") and not data.get("is_component"):
-        raise ValidationError({"is_component": [strings.Goods.FORM_NO_COMPONENT_SELECTED]})
-
-    # Validate component detail field if the answer was not 'No'
-    if data.get("is_component") and data["is_component"] not in [Component.NO, "None"]:
-        valid_components = validate_good_component_details(data)
-        if not valid_components["is_valid"]:
-            raise ValidationError({valid_components["details_field"]: [valid_components["error"]]})
-
-        data["component_details"] = data[valid_components["details_field"]]
-
-
-def validate_information_security_field(data):
-    if data.get("is_information_security_step") and data.get("uses_information_security") is None:
-        raise ValidationError(
-            {"uses_information_security": [strings.Goods.FORM_PRODUCT_DESIGNED_FOR_SECURITY_FEATURES]}
         )
